requirements.txt
pytest
playwright
allure-pytest
pyyaml
python-dotenv